requirements.txt
shodan
python-whois
bs4